pyproject.toml
[tool.poetry]
name = "get_data_to_geo_portal"
version = "0.1.0"
description = "A project to fetch and process data for a geo portal."
authors = ["Yagel Maimon <Yagelm@israntique.org.il>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
geopandas = "^0.14.3"
psycopg2-binary = ">=2.9.0"
py7zr = ">=0.20.0"
rarfile = ">=4.0"
pyogrio = "^0.11.1"

[tool.poetry.dev-dependencies]
pytest = "^8.2.0"
black = "^24.4.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
